#35 Extract expression formatting into _expressions module
Move the SBML MathML/AST -> C++ formula translation out of ChasteSbmlModel into a
new _expressions module: the AST helpers (collect/substitute names, replace the
avogadro csymbol, strip units), the infix-operator rewriter and formula_to_string
are now free functions. formula_to_string takes the model's variable-type map and
state variables as arguments rather than reaching into the model, and resolves
rateOf inline. ChasteSbmlModel keeps a thin _formula_to_string wrapper so its call
sites are unchanged.

Behaviour-neutral: reference model output is byte-identical. The infix tests move
to test_expressions.py alongside new formula_to_string tests. Trims ChasteSbmlModel
by ~350 lines.
